feat: Update README.md, docs, and script - Update README.md for cleaner output - Update docs for more info - Update stow script to prevent deleting own files

+++ script/stow.sh
@@ -45,7 +45,7 @@ for package in */; do
             target_path="$HOME/$relative_target"
 
             if [ -e "$target_path" ] || [ -L "$target_path" ]; then
-                if [ ! -L "$target_path" ]; then
+                if [ ! -L "$target_path" ] && [ ! -L "$(dirname "$target_path")" ]; then
                     echo -e "${YELLOW}[!] Overwriting conflict:${NC} $relative_target"
                     rm -f "$target_path"
                 fi
